requirements.txt
aiohappyeyeballs==2.6.1 ; python_version == "3.11"
aiohttp-cors==0.8.1 ; python_version == "3.11"
aiohttp==3.12.14 ; python_version == "3.11"
aiosignal==1.4.0 ; python_version == "3.11"
annotated-types==0.7.0 ; python_version == "3.11"
anyio==4.9.0 ; python_version == "3.11"
attrs==25.3.0 ; python_version == "3.11"
cachetools==5.5.2 ; python_version == "3.11"
certifi==2025.7.14 ; python_version == "3.11"
charset-normalizer==3.4.2 ; python_version == "3.11"
click==8.2.1 ; python_version == "3.11"
colorama==0.4.6 ; (platform_system == "Windows" or sys_platform == "win32") and python_version == "3.11"
colorful==0.5.7 ; python_version == "3.11"
contourpy==1.3.2 ; python_version == "3.11"
cycler==0.12.1 ; python_version == "3.11"
dataclasses-json==0.6.7 ; python_version == "3.11"
defusedxml==0.7.1 ; python_version == "3.11"
distlib==0.4.0 ; python_version == "3.11"
et-xmlfile==2.0.0 ; python_version == "3.11"
fastapi==0.116.1 ; python_version == "3.11"
filelock==3.18.0 ; python_version == "3.11"
fonttools==4.59.0 ; python_version == "3.11"
frozenlist==1.7.0 ; python_version == "3.11"
fsspec==2025.7.0 ; python_version == "3.11"
google-api-core==2.25.1 ; python_version == "3.11"
google-auth==2.40.3 ; python_version == "3.11"
googleapis-common-protos==1.70.0 ; python_version == "3.11"
grpcio==1.74.0 ; python_version == "3.11"
h11==0.16.0 ; python_version == "3.11"
hf-xet==1.1.5 ; python_version == "3.11" and (platform_machine == "x86_64" or platform_machine == "amd64" or platform_machine == "arm64" or platform_machine == "aarch64")
httpcore==1.0.9 ; python_version == "3.11"
httptools==0.6.4 ; python_version == "3.11"
httpx==0.28.1 ; python_version == "3.11"
huggingface-hub==0.34.0 ; python_version == "3.11"
idna==3.10 ; python_version == "3.11"
importlib-metadata==8.7.0 ; python_version == "3.11"
jinja2==3.1.6 ; python_version == "3.11"
jsonschema-specifications==2025.4.1 ; python_version == "3.11"
jsonschema==4.25.0 ; python_version == "3.11"
kiwisolver==1.4.8 ; python_version == "3.11"
loguru==0.7.3 ; python_version == "3.11"
markupsafe==3.0.2 ; python_version == "3.11"
marshmallow==3.26.1 ; python_version == "3.11"
matplotlib==3.10.3 ; python_version == "3.11"
mpmath==1.3.0 ; python_version == "3.11"
msgpack==1.1.1 ; python_version == "3.11"
multidict==6.6.3 ; python_version == "3.11"
mypy-extensions==1.1.0 ; python_version == "3.11"
networkx==3.5 ; python_version == "3.11"
numpy==1.26.4 ; python_version == "3.11"
opencensus-context==0.1.3 ; python_version == "3.11"
opencensus==0.11.4 ; python_version == "3.11"
opencv-python-headless==4.11.0.86 ; python_version == "3.11"
opencv-python==4.11.0.86 ; python_version == "3.11"
openpyxl==3.1.5 ; python_version == "3.11"
opentelemetry-api==1.35.0 ; python_version == "3.11"
opentelemetry-exporter-prometheus==0.56b0 ; python_version == "3.11"
opentelemetry-proto==1.35.0 ; python_version == "3.11"
opentelemetry-sdk==1.35.0 ; python_version == "3.11"
opentelemetry-semantic-conventions==0.56b0 ; python_version == "3.11"
packaging==25.0 ; python_version == "3.11"
pandas==2.3.1 ; python_version == "3.11"
pillow==11.3.0 ; python_version == "3.11"
platformdirs==4.3.8 ; python_version == "3.11"
prometheus-client==0.22.1 ; python_version == "3.11"
propcache==0.3.2 ; python_version == "3.11"
proto-plus==1.26.1 ; python_version == "3.11"
protobuf==6.31.1 ; python_version == "3.11"
psutil==7.0.0 ; python_version == "3.11"
py-cpuinfo==9.0.0 ; python_version == "3.11"
py-spy==0.4.0 ; python_version == "3.11"
pyasn1-modules==0.4.2 ; python_version == "3.11"
pyasn1==0.6.1 ; python_version == "3.11"
pycocotools==2.0.10 ; python_version == "3.11"
pydantic-core==2.33.2 ; python_version == "3.11"
pydantic==2.11.7 ; python_version == "3.11"
pyparsing==3.2.3 ; python_version == "3.11"
python-dateutil==2.9.0.post0 ; python_version == "3.11"
python-dotenv==1.1.1 ; python_version == "3.11"
python-multipart==0.0.20 ; python_version == "3.11"
pytz==2025.2 ; python_version == "3.11"
pyyaml==6.0.2 ; python_version == "3.11"
ray==2.48.0 ; python_version == "3.11"
referencing==0.36.2 ; python_version == "3.11"
requests==2.32.4 ; python_version == "3.11"
rpds-py==0.26.0 ; python_version == "3.11"
rsa==4.9.1 ; python_version == "3.11"
scipy==1.12.0 ; python_version == "3.11"
six==1.17.0 ; python_version == "3.11"
smart-open==7.3.0.post1 ; python_version == "3.11"
sniffio==1.3.1 ; python_version == "3.11"
starlette==0.47.2 ; python_version == "3.11"
supervision==0.26.1 ; python_version == "3.11"
sympy==1.14.0 ; python_version == "3.11"
torch @ https://download.pytorch.org/whl/cpu/torch-2.2.2%2Bcpu-cp311-cp311-linux_x86_64.whl ; python_version == "3.11"
torchvision @ https://download.pytorch.org/whl/cpu/torchvision-0.17.2%2Bcpu-cp311-cp311-linux_x86_64.whl ; python_version == "3.11"
tqdm==4.67.1 ; python_version == "3.11"
typing-extensions==4.14.1 ; python_version == "3.11"
typing-inspect==0.9.0 ; python_version == "3.11"
typing-inspection==0.4.1 ; python_version == "3.11"
tzdata==2025.2 ; python_version == "3.11"
ultralytics-thop==2.0.14 ; python_version == "3.11"
ultralytics==8.3.169 ; python_version == "3.11"
urllib3==2.5.0 ; python_version == "3.11"
uvicorn==0.35.0 ; python_version == "3.11"
uvloop==0.21.0 ; sys_platform != "win32" and sys_platform != "cygwin" and platform_python_implementation != "PyPy" and python_version == "3.11"
virtualenv==20.32.0 ; python_version == "3.11"
watchfiles==1.1.0 ; python_version == "3.11"
websockets==15.0.1 ; python_version == "3.11"
win32-setctime==1.2.0 ; python_version == "3.11" and sys_platform == "win32"
wrapt==1.17.2 ; python_version == "3.11"
yarl==1.20.1 ; python_version == "3.11"
zipp==3.23.0 ; python_version == "3.11"
